debian: Add suffix -lx
The packages are not identical to the upstream debian libvirt
packages. They have additional r4d hypervisor support. When updating a
system, packages with higher version number are taken. To ensure that
the packages with r4d support are not replaced by upstream packages
with higher version number, add -lx to the package name. A "provides"
with the origin package name is added to be compatible and prevent
dependency problems.

The resulting packages are now:
libvirt-clients-lx
libvirt-daemon-lx
libvirt-daemon-system-lx
libvirt0-lx
libvirt-lx-doc
libvirt-lx-dev
libvirt-sanlock-lx
libnss-libvirt-lx

Signed-off-by: Martin Kaistra <martin.kaistra@linutronix.de>
Signed-off-by: Anna-Maria Gleixner <anna-maria@linutronix.de>