Move data files from libvirt-daemon to libvirt0
These files are used internally by the library, so they
should be shipped along with it rather than with the daemon.
This is consistent with the upstream libvirt.spec file.

The pattern is partially expanded in the libvirt0.install
file to avoid having to remove a specific subset of data
files later on as part of debian/rules.