d/rules: strip symbolic-functions linker option
This is only set by-default on Ubuntu today.
It mostly works, but as the word mostly implies some cases are different.
Libvirt now became one of them.

As discussed DEB_flag_MAINT_STRIP already is conditional therefore no
need to vendor the check, but better add a comment why/what it does.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>