enhance API check
libvirt perform an API check during test to verify the libvirt driver
naming convention i.e. hypervisor in lowercase plus function in Camelcase
starting with a capital letter.

The hypervisor r4d does not pass the test due to the numeral.
Enhance the Api test to allow numerals in the hypervisor name.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>