Handle domain snapshot callbacks
Jenkins libvirt-slave-plugin calls virDomainSnapshotNum() and
virDomainListSnapshots(). R4D does not provide snapshots and initialize
the corresponding callback pointer with NULL and let libvirt handle the
whole thing. Somewhere in the call stack the error result got lost and
the error value -1 is interpreted as index.

Implement the missing functions as a workaround to avoid nasty
backtraces.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>