Reorganize documentation
The upstream project decided to provide a README.md instead of a
README.adoc file. The README.md file is also shown markup rendered on
github. Therefore it is not needed to add a README.adoc any more.

Integrate the content of README.adoc in README.md and drop all references
to README.adoc.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>