No acls for now
libvirt is capable to handle stateful drivers i.e. the state of a
system is cached and handled internally by libvirt. As a side effect ACLs
are implemented and handled internally. R4D is not capable to be handled
by a stateful driver due to its full asynchronous design.

Regrettably the former implementation of the R4D driver faultily claims
to be a stateful driver. Change the driver to be stateless and remove
all calls to stateful ACL checks.

Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>